moved image-sizes under `image` dir
